More sane package structure
